app/pyproject.toml
[tool.poetry]
name = "team-ai"
version = "0.1.0"
description = ""
authors = ["Team-aide incubation <team-aide@thoughtworks.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
Authlib = "^1.3.0"
boto3 = "^1.34.82"
faiss-cpu = "^1.8.0"
fastapi = "0.110.2"
google-cloud-aiplatform = "1.44.0"
gradio = "4.25"
itsdangerous = "^2.1.2"
Jinja2 = "3.1.3"
langchain = "0.1.11"
langchain-google-genai = "^1.0.1"
langchain-google-vertexai = "^0.1.2"
loguru = "^0.7.2"
ollama = "^0.1.8"
openai = "^1.17.0"
pyarrow = "^15.0.2"
pypdf = "^4.2.0"
python-frontmatter = "^1.1.0"
shapely = "^2.0.3"
tiktoken = "^0.6.0"
typing-extensions = "^4.11.0"
typing-inspect = "^0.9.0"
unstructured = "^0.13.2"
uvicorn = "^0.29.0"
langchain-openai = "^0.1.2"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-mock = "^3.14.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
omit = ["tests/*"]